change tests to reflect new deliverables
