+++ test/1-MovieCard.test.js
@@ -10,21 +10,12 @@ import MovieCard from '../src/card-components/MovieCard'
 
 
 describe('<MovieCard />', () => {
-  
+
   it("assigns a default prop of 'Unknown' for the title prop", () => {
     expect(MovieCard.defaultProps.title.toLowerCase()).to.equal('unknown')
   })
-  
-  it("assigns a default prop of null for the IMDBRating prop", () => {
-    expect(MovieCard.defaultProps.IMDBRating).to.equal(null)
-  })
-  
+
   it("assigns a default prop of ['No Genre(s) Found'] for the genres prop", () => {
     expect(JSON.stringify(MovieCard.defaultProps.genres)).to.equal(JSON.stringify(['No Genre(s) Found']))
   })
-  
-  it("assigns a default prop of 'default' for the poster prop", () => {
-    expect(MovieCard.defaultProps.poster.toLowerCase()).to.equal('default')
-  })
-  
 })
