+++ test/2-CardFront.test.js
@@ -8,11 +8,24 @@ configure({ adapter: new Adapter() })
 
 import CardFront from '../src/card-components/CardFront'
 
+
 describe('<CardFront />', () => {
-  
-  it('uses the poster prop as a background image, rendering only the necessary HTML', () => {
-    const cardFront = shallow(<CardFront poster="you spin me right round"/>)
-    expect(cardFront.html()).to.equal('<div class="card-front" style="background-image:url(you spin me right round)"></div>')
+
+  const genres = [
+    "Never gonna give you up",
+    "Never gonna let you down",
+    "Never gonna run around and desert you"
+  ]
+
+  const cardBack = shallow(<CardFront title="Rolling in the Deep" genres={genres} />)
+  const text = cardBack.text()
+
+  it("correctly renders the title prop", () => {
+    expect(text).to.contain('Rolling in the Deep')
+  })
+
+  it("correctly renders the genres prop as comma seperated strings", () => {
+    expect(text).to.contain(genres.join(', '))
   })
 
 })
